requirements.txt
more-itertools==8.13.0
Pillow==9.2.0
